Allow using parts as a library outside of Envoy
